Partially flesh out the readme
Rename readme from "README.md" to "README.rst". reStructuredText is the standard
markup format for Python projects, and this project is primarily Python.

The readme is not fully fleshed out. However, it contains much useful
information, and it should be enough to get a developer up and running with the
project in its current state.