Flesh out readme. Add lighttpd configs.
Thorougly flesh out the readme. The most interesting changes occurred in the
"Production Setup" section of the "Deployment Guide", but plenty of changes were
made to other portions of the document.

The lighttpd configs for this application are similar to the lighttpd configs
for ELTS. However, they assume the use of the gunicorn HTTP app server, not the
flup SCGI app server.