Restrict who can update or delete characters.
If a user owns a characer, or if a user owns a campaign to which that character
belongs, the user can do the following:
* update the character
* delete the character
* fetch the character update form
* fetch the character delete form

If a user is denied while fetching an update or delete form, an HTTP 403
response is returned. However, no nice HTML response is returned. Additionally,
no unit tests have been written for the functionality added in this commit.