Fix links in character/id/ HTML template
If a user has the rights to view a character, but does not own that character,
the user does not have the rights to view the character's possessions, hit
locations, etc. Update the HTML template for path `character/id/` so links to
that character's possessions, hit locations, etc. if appropriate.

This fixes issue #22.